requirements.txt
python-dotenv
langchain-huggingface
